refactor: move thread event buffer owner
